Update patch set 5
Patch Set 5:

the numbers ones are run using the art/test/run-test script and can be run using --dalvik or --jvm for comparison.

the non-numbered ones were what we started with in art. where they duplicated or were able to be moved to "run-tests", we did so. however, the JniTest uses a .so and run-test relies mostly on output diffing, so some of the threaded tests have remained.

Patch-set: 5
Reviewer: Gerrit User 1003723 <1003723@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0