Update patch set 1
Patch Set 1:

You'll probably need to add a new command-line dumping option for this to work properly.  In the existing --dump-timings, each phase shares a relatively consistent share of the timing logger overhead, so the overhead can largely be ignored for the purposes of comparing the relative cost of the phases.  However, with this CL, the compiling phase is not burdened with the (pretty significant) cost of timing logging - which adds to overall compile score (and warps it pretty substantially).
For example, on my desktop with the existing logging scheme the "Compile Dex File" phase takes 2.97 seconds.  With this CL, the added overhead of logging the splits of the compiler's passes sends the cost to 3.5 seconds.
For this reason, I typically have been doing two different runs - one to get relative phase costs, and another to collect compiler pass costs.

Patch-set: 1