Update patch set 3
Patch Set 3:

Considering how slow we're going to be with a full deopt, I think we should compute it dynamically (I guess we can already compute it today?). The likelihood of having 1) jdwp attached, 2) usage of breakpoints by the user and 3) setting a breakpoin on an inlined method, makes me think that adding extra bytes in the aot for carrying the reverse mapping is not worth it.

Patch-set: 3
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0