Update patch set 1
Patch Set 1:

Well they are used for the logging. The original code did:

TimeSplit(MIROpt:Name);
CallPassFunction();

Now I moved things and moved the Name into the Pass and put the MIROpt outside. If someone is postprocessing the logger, it would be logical that changing the name as this patch does will mess that up.

Actually, the other simple solution is to change the const char const *name in TimingLogger to std::string name as it is done in CumulativeLogger.

Though Ian worried about std::string allocation, we aren't creating thousands of loggers it would seem, this is only done if we are asked to log.

Patch-set: 1