Update patch set 1
Patch Set 1: Code-Review+1

(5 comments)

Nice, and there is obviously a lot of other places we can use this, as well as cleaning up the existing launch pads to use this mechanism. One thing to think about is the delta encoding in the pc mapping tables. It could be we see less incrementally increasing pcs following this change.

Patch-set: 1
Label: Code-Review=+1