Update patch set 2
Patch Set 2:

The motivation for the access check rewrite is clear: have the access check in one place instead of 2 (methods) or 3 (fields). Or more after I add another access check for fields referenced from an inlined function rather than the current compilation unit.

The motivation for getting rid of the helper functions in the compiler driver is that they don't make much sense. Just look at the dex_cache usage from ComputeInstanceFieldInfo(). A SirtRef is temporarily acquired in ComputeFieldReferencedFromCompilingMethod(), then a possibly wrong one directly in the ComputeInstanceFieldInfo() and it can be corrected in ComputeCompilingMethodsClass()?! Complete mess.

Patch-set: 2