Update patch set 2
Patch Set 2:

Performance: Are you trying to tell me that the mere _existence_ of a SirtRef is costly? I can imagine it costs something when the occasional GC kicks in but otherwise it should not cost anything. As to construction and destruction of SirtRefs, this CL is reducing that.

I don't see why access checks should differ between compile time and runtime (as long as we can actually resolve the types at compile time).

Patch-set: 2