Update patch set 1
Patch Set 1:

(2 comments)

> About:
 > 
 > - We check whether transaction mode is enabled or not in each
 > field/array write
 > routine to not miss anything. This may have negative impact on
 > performance.
 > 
 > I guess you're talking about runtime performance when running ART
 > in production? How could it not have a negative impact?

It certainly has negative impact as is because it's pretty invasive right now. I'm planning on reducing this. On the other hand, compiled code does not use these routines on the fast path which represents most of the cases I think where it should not hurt. It more concerns modifications to the runtime internal state at the moment that I'm planning to work on.

Patch-set: 1
Label: Verified=0