Update patch set 2
Patch Set 2:

(1 comment)

I don't understand the motivation for the CL. The locking should only exist in the stats in debug builds, no?
In general dex_compilation_unit is on my "to be deleted" list. It was a way for a lazy engineer to avoid passing the appropriate information to methods and to just pass around a big aggregate. The aggregate itself has little meaning and questionable correctness.

Patch-set: 2