Update patch set 1
Patch Set 1:

So rather than have a "GC running", albeit trimming, but for that GC to have type "no GC" I'd rather we added the enum entry and dealt with the ramifications. Other wise I can imagine code that uses the "no GC" enum value to mean that no GC is running. I will also be confused looking at the state of the heap in a debugger.

Patch-set: 1