Update patch set 1
Patch Set 1:

(2 comments)

For iget I don't see why doing a pre-pass is going to benefit us, it adds storage :-/
For iput a naive peephole pass would see writes to the same class which could allow the write barrier to be elided. In any case the pertinent information of object and type isn't carried in the annotation. Similarly for sput/sget, where the hope is to elide initialization checks in the case they of a class repeatedly reference fields in a different class. If we can detect this latter case then perhaps we want to look to promote the class to a register. I'm not sure its dynamically interesting though.
In summary, is all of this actually going to buy anything meaningful? Perhaps the statics case is going to be more meaningful with inlining, but your annotations aren't full field references as they skip the defining dex file. As is I'm seeing lots of cost for what is probably little gain.

Patch-set: 1
Label: Verified=0