Update patch set 2
Patch Set 2: Code-Review-1

There's not enough information in this change to determine whether its a good or bad change. In particular its paying a cost that its unlikely to recoup. I'd prefer to see it as part of a larger patch set so that we can work out whether we're going in the right direction or not.

Patch-set: 2
Label: Code-Review=-1