Update patch set 3
Patch Set 3:

Hi Hiroshi,

I'm totally fine with your patch. I will check whether the issue is reproducible or not however the reproducer is not stable, so if I meet this failure again I will come back with this issue. 

The reproducer itself is pretty complex and based on Harmony Reliability Tests, so I doubt that it is possible to create any simple reproducer for this issue to contribute. Most of the root cause was found by source code analysis.

Patch-set: 3
Reviewer: Gerrit User 1040038 <1040038@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0