Update patch set 1
Patch Set 1:

we have _test.cc gtests around the art tree that are part of the test-art make target. we have other java based tests in art/tests. this type of bug is usually found with one of the heap verification modes. I don't think we have a gtest that is a good starting point to regress this type of issue, which was found mostly with monkey and user testing

Patch-set: 1