Update patch set 3
Patch Set 3:

when we build the *-dalvik-host branches, it uses dalvik-host as the target, so it doesn't bother to try to build an image. 

you can look at the manifest for those branches to see what subset of projects is needed to build those targets. I also set the environment variables WITHOUT_LIBCOMPILER_RT=true WITHOUT_CLANG=true to avoid dependencies for dalvik, but ART might need some of those.

Patch-set: 3