Update patch set 5
Patch Set 5:

Fwiw, I'm not asking you to make this a large CL but to finish the WIP and then provide data on whether the change pays off. As noted earlier your annotations are incorrect as they are not correct references, the rely on the index being local to the current dex file where this needn't be the case following inlining. Probably the most interesting case of the optimizations flowing from this work firing.

Patch-set: 5