Update patch set 4
Patch Set 4: Code-Review+2 Verified+1

Self approving revert of change that didn't have +2.
The original change was +2ed in an earlier form but I'd put a -1 (I'd prefer for this not to be merged) review very shortly after.
The change author then try to conflate the issue with dependent CLs being merged afterwards. This change disables that code and introduces TODOs.
To be specific as to what the issue with the original change is, we have an agreement between the verifier, compiler interface and runtime slow paths as to what the correct behavior for dex instructions is. The original change moved parts of that agreement out of the compiler interface into a new notion of "annotation", the construction of which sidelined the compiler interface. Without an interface it makes it difficult to reason about the correctness of operations as the interface now exists in multiple places. Its a design goal that this should exist in a single place. This doesn't mean that "annotations" can't be created, they can in the MIR code. However, when it comes to the assumptions they are making they need to query the compiler interface. If the compiler interface is inappropriate or needs adapting then so be it. If the compiler interface requires performance tuning then its also fine to use -inl.h files. The point is we shouldn't be dispersing the control of when we go fast or slow as its a cornerstone of the correctness of our AOT system.

Patch-set: 4
Reviewer: Gerrit User 1010118 <1010118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2
Label: Verified=+1