Update patch set 2
Patch Set 2:

Let's say no to this one for now.  This code base originated in the Dalvik world and as Ian points out, it doesn't quite fit in with the rest of ART.  However, I'd rather address this with a dedicated house-cleaning effort which identifies a target structure and works toward it rather than piece-meal efforts.  I don't consider right now to be a good time for a major restructuring, but let's revisit when things calm down.  Meanwhile, we can live with a little extra bloat in mir_graph.cc.

Patch-set: 2