Update patch set 3
Patch Set 3:

Vitaly, if you are still reading this, I'm having a problem running the harmony reliability test. I followed the README but encountered an error that  ../../../scripts/properties.xml isn't found when I do "ant -f run.xml". The file does not exist as it's outside of the directory subtree. I'm not familiar with ant or its xml language. Is there a working command sequence starting with the "svn co ..."? Thanks.

Patch-set: 3