Update patch set 1
Patch Set 1:

Thanks for the explanations, I agree that makes sense.

It looks like Thread Sanitizer is now part of Clang 3.4 and GCC 4.8 (see [1]). Could this be considered a valid alternative for thread-safety issues if we update the host toolchain?

I'm not sure I understand what -Wthread-safety currently does in GCC 4.6, but it looks like using thread sanitizer requires adding '-fsanitize=thread -fPIE' at compile time, and '-fsanitize=thread -pie' at link time. However, the main difference is that this performs runtime checks (and thus impact performance), instead of compile-time ones, so this should probably be restricted to building a special version of ART to debug these specific issues.

Another option is to keep -Wthread-safety, but force the host build to use Clang which supports it, as far as I understand.

[1] https://code.google.com/p/thread-sanitizer/wiki/CppManual

Patch-set: 1