Update patch set 1
Patch Set 1:

Thread sanitizer is distinct from -Wthread-safety that is also known as annotalysis. One is dynamic analysis of race conditions and the like, the other is annotations on code that are statically analyzed to ensure locks are handled with care, acquired in an ordered manner, etc. My requirement is just to get at least one build with thread safety checking, either from clang or GCC, otherwise our use of it will bit rot.

Patch-set: 1