Update patch set 1
Patch Set 1:

Would it be possible to put a snippet of code in:
art/test/083-compiler-regressions/src/Main.java
the code should do something like the mulBy1Test which prints a message on whether the operation gave an expected or unexpected answer.

The file:
art/test/083-compiler-regressions/expected.txt
should be updated to ensure the output of the test matches expectations.

Patch-set: 1