Update patch set 1
Patch Set 1:

Sorry, I rushed to put a -2 on this because Ian had +2d.

So my main desired is that the semantics of requesting an addr match between the different functions. It seems like it should not be fatal to have a failure.

I also don't think MAP_FIXED should be supplied automatically, again its not in the other case. If the caller wants that, it should pass it. That will cause MAP_FAILED, then we can notify the caller via a NULL result and *error_msg.

Finally, if possible, we should expand the mem_map_test to cover changes. In particular, it would seem easy to add a test that does a MapAnonymous to get a region, then does it again at the same address without MAP_FIXED, to see that it works but returns a different address, and with MAP_FIXED to make sure it fails.

Patch-set: 1