Update patch set 2
Patch Set 2:

Note that IIUC, all the existing call sites to MapAnonymous() with a non-null requested address do not check that the actual address and the requested address match (it only checks the returned mem map is not null.) Hence, the motivation for PS1.

The suggested test strategy didn't work as attempting to map at an already-mapped address is fatal due to CheckMapRequest().

Patch-set: 2
Label: Verified=0