Update patch set 2
Patch Set 2:

> Note that IIUC, all the existing call sites to MapAnonymous() with
 > a non-null requested address do not check that the actual address
 > and the requested address match (it only checks the returned mem
 > map is not null.) Hence, the motivation for PS1.

I didn't realize we didn't expose the flags argument for MapAnonymous like we did for MapFileAtAddress. That is what I was getting at when I said "If the caller wants that, it should pass it." But its clear from having to change all the call sites that this is not the case...

Having looked at this, its clear that MapAnonymous is not trying to be as general purpose as the other methods. So I would go back to the first strategy of supplying MAP_FIXED if addr != nullptr. However, we shouldn't be failing with a CHECK in these basic library routines if things go wrong if we've already done the error_msg plumbing.

 > The suggested test strategy didn't work as attempting to map at an
 > already-mapped address is fatal due to CheckMapRequest().

well, I'd have said we shouldn't be doing the CheckMapRequest when map_at_exact_addr == true. I guess that still isn't enough. It seems like CheckMapRequest should just report via error_msg to be consistent with the new style. For MapAnonymous, I would only do the CheckMapRequest if the MAP_FAILED, since you are using MAP_FIXED. For MapFileAtAddress, I would only do it "if (!reuse && kIsDebugBuild)"?

Patch-set: 2