Update patch set 5
Patch Set 5:

> Sorry, I misread your map_fixed man page snipped backwards. since I'm just
> confusing things, I took a stab at this myself.
>
> basically my point is we don't really want to do the slow path
> CheckMapRequest ever. We just want to ask the kernel to mmap, if we don't
> get what we want, then we use the CheckMapRequest to report the issue.
>
> this seems to make sense for the MapFileAtAddress case too. making those
> aligned allows me to push the actual vs expected into CheckMapRequest.
>
> Does this meet the original goals of the change still?

The original goal is to explicitly fail in MapAnonymous() when expected != nullptr and expected != actual, because call sites don't check this and that could lead hard-to-diagnose bugs. It doesn't seem like PS5 achieves this because it does not pass in MAP_FIXED and CheckMapRequest() does not fail when expected != actual as long as there's no overlap issue (as an aside, an overlap issue isn't likely to happen as we aren't passing in MAP_FIXED.)

Re: MapFileAtAddress, where it's likely the caller supplies MAP_FIXED, it seems that the overlap check comes after the mmap call, which may clobber existing mapped regions and mask the overlap issue, according to the mmap man page snippet I quoted. Is this OK?

Patch-set: 5
Label: Verified=0