Update patch set 4
Patch Set 4:

This gets as far as running the interpreter, but bailing out in an exception.
The lib directory should have the 64bit binaries copied in for this to work.

There is work outstanding to be done on the JNI compiler.

external/llvm/llvm.mk needs arm64 added to it.

Patch-set: 4
Label: Verified=0