Update patch set 5
Patch Set 5:

(2 comments)

I really don't want to use map_fixed, it means you are have to be careful and check for overlap in the common case.

I'm aware of the other patch, I think we are still discussion about other options. Even for that approach, I think we need to do something a little better. I was hoping to get this in first then address that one.

Patch-set: 5