Update patch set 1
Patch Set 1:

Yea, it was the LIR mem consumption that I was considering addressing last year.  If I recall, I estimated a LIR memory saving of about 30% if we converted the operand array in LIR to be a flexible array and allocated only the memory needed for each LIR kind.  I didn't proceed at the time because I didn't have data on memory consumption for the verifier and the rest of dex2oat.  There are a few places where we change LIR opcodes, which would be a bit messy in the flexible array scheme, and I didn't want to go down that road until I was sure it would make a system-wide difference.

Patch-set: 1