Update patch set 2
Patch Set 2:

my point is that the condition above leads to using memcpy invocation with overlapped arrays. This is prohibited by memcpy specification. memcpy does not specify whether it uses forward or backward copying. For not-overlapped arrays both are safe to use. So memcpy function is allowed to use backward copying what is wrong for above condition. From other side memmove is safe in terms of overlapped arrays and it can be used.

Patch-set: 2