Update patch set 1
Patch Set 1:

I'm wondering if we should make it really explicit and have a wrapper for GenBarrier specifically for terminating OpIT blocks?

void GenTerminateITBlock() {
   GenBarrier();
}

or something like that. Then it would be obvious to someone writing a new IT block that they should terminate it rather than the rather obscure and generic GenBarrier()

Patch-set: 1
Label: Verified=0