Update patch set 1
Patch Set 1: Code-Review-1

There's a bunch going on here:
1) a review comment about a lock ranking issue for having quasi atomic while building the boot image
2) disabling the use of atomics on 32-bit systems and pushing them into using locks
3) a problem with volatile 64bit load/stores allowing tearing in C++

We shouldn't merge this change as (2) is a performance regression.
For (1) we should address this for architectures that need quasi-atomic, namely mips32 where currently we've not seen the lock ranking issue.
For (3) we use "volatile uint64_t*" and document that we don't expect tearing. From your bug it seems the compiler is generating load/stores that allow tearing. The correct solution here is to use std::atomic, however, its a WIP to get libc++ support in ART. A work-around is to use inline assembly and say 64-bit XMM registers on x86. Alternatively there may be a language extension to use rather than the volatile 64-bit loads and stores.

Patch-set: 1
Reviewer: Gerrit User 1010118 <1010118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1