Update patch set 2
Patch Set 2:

Andreas, Ian, the issue is not in JNI compiler, thus there is nothing to add to jni_compiler_test. It is pure interpreter issue with generic JNI.
The root-cause is an incorrect parameter order passed on stack by quick entrypoints. At this moment it is clear that parameter passing should be revised in entrypoints, trampoline, jni compiler and QCG, but before that the jni compiler should be re-enabled in AOSP, since it was broken by thread offsets patch.
I would suggest to merge this patch to aosp (probably, as revertme) in order to unblock further 64-bit work in aosp master tree.

Patch-set: 2