Update patch set 1
Patch Set 1:

(1) I guess you did not see the issue with mips32 due to low parallelism on the system where you used it. The fact that on x86 host mode with quasi atomic the lock ranking issue appears says that the problem exists. But as soon as I do not know how to fix it I'm leaving this issue as is.
(2) I see your point let's try to avoid this.
(3) The problem is that usage of 64-bit XMM register to reach atomic behavior is possible only if the address is 8-bytes aligned. Are we sure that this always true? If not we need to find other workaround...

Please note that we will need to do the same thing in ART quick compiler. Currently it does not support atomic long/double store/load due to the same issue and to use xmm there we need to ensure 8-byte alignment there as well...

I will try to come up with some solution...

Patch-set: 1