Update patch set 10
Patch Set 10:

You may have a point there.  Because the deduplication is done by comparing the binary code before the relocations are applied it is indeed possible that two sequences that exactly the same except for the helper calls will be wrongly deduplicated.

Thanks.  I'll deal with that by putting the helper offset into the instruction prior to relocation.

Nice catch.

Patch-set: 10