Update patch set 2
Patch Set 2: Verified+1

So running the ReflectionBenchmark from libcore, it appears using std::string hurts the performance of invoke by 25-30% (tested on N4). I decided to get back to using StringPiece and creating ClassHelper with a larger scope. Doing so does not seem to impact the performance.

Patch-set: 2
Label: Verified=+1