Update patch set 1
Patch Set 1:

I understand that for the image. But the instruction set does not make sense. The runtime is compiled for exactly *one* instruction set. With your change, I could pass "mips" to a runtime executing "x86." There is no logic involved in finding out which ISA the runtime was compiled for (in the runtime itself), so there's no duplicate logic (there should only be "logic" in the app_process).

Patch-set: 1