Update patch set 2
Patch Set 2: Code-Review+1

Big patch and I apologize.

In order to get the Loop formation system up, we added a lot of infrastructure in the ME. This patch is one of them that adds abstraction of expressions for the MIRs.

There is also a bit of changes in the DecodedInstructions and in the rest of the ME.

There are conflicts with other of my patches but I'm uploading to get feedback. Once we figure out what goes forward, I will create a chain instead and remove the conflicts.

Patch-set: 2
Label: Code-Review=+1