Update patch set 2
Patch Set 2:

A "big fd" is a file descriptor which is so big (e.g. >= 800), that a normal process can never (or rarely) have it open.

Using Thread method to get a stack trace can work if the Java process monitors its file descriptor leaks by itself. But it depends on the assumption that the process keeps running even when it already has 1024 open file descriptors. However, a process often crashes in this case. So it's more reliable to record the big FDs & corresponding backtraces in a separate daemon process.

Patch-set: 2