Update patch set 2
Patch Set 2:

Our intent is to remove DecodedInstruction, its part of the MIR so in the short-term we should just move it into the quick compiler. Part of the problem with DecodedInstruction is the multiple decodings of vB. Instruction is less error prone. There is a single use of DecodedInstruction in the verifier so I'll remove that and move DecodedInstruction to MIR.

Patch-set: 2