Update patch set 3
Patch Set 3:

> How much multi-threaded do we want this? 

I assumed we would try to parallelize at the same granularity of CompilerDriver for the setup, which is by Class I believe.

 > We could just give each
 > InitMapMethodVisitor its own ArenaStack and have them run
 > concurrently and modify the InitCodeMethodVisitor's fixup. Or we
 > could move the fixup to its own pass and run the InitCodeMethodVisitor
 > (again with its own ArenaStack) alongside the three
 > InitMapMethodVisitors.
 > 
 > Do you envision a finer-grained work split? In that case I would
 > drop the OatWriter part of this CL and consider the rest a simple
 > cleanup.
 > 
 > And do you really want the writing passes to be multi-threaded? Do
 > you want to split the OutputStream?

Ian and I weren't sure if the parallel writing would offer any improvement or not. I was going to try mmaping the target file and then writing the memory in parallel or dup(2)ing the fd and writing to different chunks in parallel with a test program to see if it would buy anything.

Patch-set: 3
Reviewer: Gerrit User 1003723 <1003723@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0