Update patch set 1
Patch Set 1:

Ian, the problem here is not related with sign extension. The main problem is MAP_32BIT on x86_64 just could have 2G memory range (0~ 0x7ffffffff). So if the expected_ptr is 0x70000000 and size is larger than 256M, the allocation could fail.

So we should try best not to use MAP_32BIT. If the expected_ptr is not NULL, we shouldn't use MAP_32BIT.

Patch-set: 1