Update patch set 5
Patch Set 5:

I'm not clear on a good name as its not clear what you are trying to do. You are trying to get information from the compiler on points the verifier has flagged as of interest? Why doesn't this change make the method verifier generate table information for every dex pc? The change just seems to undo some optimizations to eliminate dead code and remove redundant meta-data.

Patch-set: 5