Update patch set 1
Patch Set 1: -Code-Review

The idea was to pull the infrastructure of the pass driver, which is previously only used in the ME, out of the ME and generalize it so that any part of AOSP could use it.

This patch contains of the combination of 1) a general pass driver infrastructure which is in runtime and 2) an updated ME pass driver which derives from number 1 which continues to live in compiler/dex.

Now, about the static analysis:
Previously in https://android-review.googlesource.com/#/c/91803/ I used a pass driver like infrastructure but there was repetitive code from the ME pass driver. This patch solves that. I am working on a version that will update patch 91803 which will make it dependent upon part 1 (the generalized part) and not part two (which is the ME pass driver used by quick).

Patch-set: 1
Label: Code-Review=0