Update patch set 1
Patch Set 1: Code-Review+1

This change is part of x86_64 QCG functional enabling, thus I would suggest to merge it.

Andreas is totally right about overhead and the need of the same change for 32-bit, but it is optimization work, not functional enabling.

Thus we should have a separate optimization task for 32-bit first. Then we should do such task for 64-bits.

Such optimization work is possible only after ART QCG will become fully functional in 64-bit, which is blocked, particularly, by merging of this patch.

Andreas, is it acceptable to merge this patch as part of functional enabling (equal to 32-bit) and work on optimization as a separate task?

Patch-set: 1
Reviewer: Gerrit User 1039576 <1039576@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1