Update patch set 2
Patch Set 2:

> Why is this being done at runtime and not at compile time?
 > (assuming the classes were available at compile time)

Ian and I had no real memory of this place. I'm guessing when it was done was based on Dalvik, but not sure it would be disallowed if done earlier. it is possible that somehow this is expected to fail only at class init time in the RI.

Patch-set: 2
Reviewer: Gerrit User 1003723 <1003723@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0