Update patch set 2
Patch Set 2:

I put in a ScopedArenaAllocator for the usedef. Since the allocated data is live through calls to the use-def creation, I put the arena in the mir_graph.

I have one patch that makes a cleanup pass driver that is getting ready to upstream. Once that one is up, I will add the pass to actually call this code and create the use-def chains.

The only major comment you had was that you wanted the defs in the mir_graph. I still think that it is not an automatically good idea. Do you wish that I do it anyway?

Patch-set: 2