Update patch set 2
Patch Set 2:

(1 comment)

Sorry, I don't understand why you need to add this check. IMO, it should be handled in ThreadList::SuspendThreadByThreadId where we look for the corresponding thread in the thread list (on first loop iteration). Could you please clarify what is the issue and assertion check you hit? 

What is the jpda test failure you mention? Is it an existing test in AOSP? If not, it would be nice to also add a test here.

Thanks!

Patch-set: 2