Update patch set 1
Patch Set 1: Code-Review-1

I'm a little concerned about this.  It's been a while since I looked at but I do seem to remember races on shutdown and that's why the check for !ShuttingDown() was there.  I can't remember exactly what the issue was though.

So what is the problem we are trying to fix here?  Worst case is that we miss the last profile run before the process shuts down, but do we really need to have a profile captured during shutdown?  Do we care what methods take time at shutdown?

This will be hard to test I think.

Patch-set: 1
Label: Code-Review=-1