Update patch set 1
Patch Set 1:

For long running apps I agree with Dave that's probably not a big deal is we loose the last batch of profile data. And it's not even sure that there's anything there (it's a high chance there isn't due to the exponential backoff...).

For short execution we may actually miss any kind of profiling (I can think of a number of apps I use and exhibit this pattern).

I can also see however how this will slowdown any debugging and experimentation (I had similar issues).

If the race is not very hard to fix we should probably try do it. If you remember its source I can look into it.

Patch-set: 1