Update patch set 2
Patch Set 2:

The race was found during development and, if I remember correctly, happened when shutting down the process.  If you can try running the program using dalvikvm and if it shuts down correctly then it all should be OK.  I was just worried about removing a constraint that was there for a reason, but if there is a need to have the last 20 seconds of a processes's runtime then if we test it and it all looks good, then I'm ok with it.

Patch-set: 2
Reviewer: Gerrit User 1036486 <1036486@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0