Update patch set 1
Patch Set 1:

The initial though for this was to keep the "compile" time as close to zero as possible at the expense of runtime.  This is still an interpreter so the performance is that of an interpreter (except about 3 times as fast with the wind behind).  The goal was to eliminate the 'decode' and 'fetch next' phases of the interpreter cycle, while retaining the basic semantics of shadow frames, vregs, etc.  

Since this is definitely a WIP I am open to any suggestions to make life easier.  The majority of the work was writing the xlate.S file (the hand coded ARM sequences), but I am not sure that using a programatic assembler would be any easier than using the source one.  A programatic assembler would also be slower to execute since all the binary conversion would have to be done inline and not ahead of time.

Patch-set: 1