Update patch set 1
Patch Set 1: Code-Review+1

I believe a change similar to the one in gen_invoke.cc (in this patch) must be applied to the FlushIns version in target_arm64.cc, but - as Andreas' patch is now merged - it makes sense to merge this as well. Changes to the arm64 backend can go in a separate commit.

Patch-set: 1
Reviewer: Gerrit User 1037505 <1037505@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1