Update patch set 3
Patch Set 3:

Since suspend checks serve as optimization barriers, I wonder if it makes sense to move this concept entirely to the middle-end by adding an extended MIR for the check semantics. Because right now any ME optimizations would have to be aware on whether the BE will generate a check. Additionally, moving it to ME makes sense because even the elision works by doing analysis on the ME graph.

Patch-set: 3