Update patch set 1
Patch Set 1:

(1 comment)

Thanks.

Based on what we discussed, it seems like a pure JNI (JNI refs) solution won't work for the debugger's needs and we need to keep some sort of special handling for the debugger like we do now.

Here's one more idea: Change the map key to the object's identity hash code and store Object* as part of the value type. As identity hash codes aren't completely unique, we'd need to use a multimap where after we look up with an identity hash code key, we need to iterate over multiple value entries and find the right value entry based on the equality of the Object* pointer. Since we don't store Object* as the keys and the identity hash codes won't change, we won't have to rebuild the entire map and we can insert the read barrier code at the time of access to individual Object*. I think intern_table does something similar for strings.

Would this work?

If not, I might punt on this for now.

Patch-set: 1