Update patch set 1
Patch Set 1:

I think the main cost is having to install the identity hash code in the lock word, which puts the object in a state where locking on the object becomes more expensive. Note this cost incurs only when a debugger is attached, meaning during development time (and only for objects that the debugger tracks.) As such, I don't think the cost is the major concern here. More to do with whether it works or not.

Patch-set: 1