Update patch set 1
Patch Set 1:

A few more thoughts:

If we use the identity hash code as the map key, since the map value already holds a jobject, we won't need to store Object* in the map value. And, we won't likely have GC to call into the debugger as part of the 'sweep system weaks' call as the map key/comparator are stable and there's no raw Object* stored in the debugger.

JobjectComparator internally calls IdentityHashCode(). Not always, but likely often. If the cost of taking identity hash codes is a concern, the two approaches may not be much different in this regard.

If we use JobjectComparator as the map key, we'd have to temporarily disable (block) accesses to the map while the GC sweeps JNI weak globals and update/clear them, at which point we'd need to rebuild/update the map, as the map may be invalid when weak globals are cleared (a mismatch between the comparator and the structure of the map). I guess this is what Ian meant by "fix the race with sweeping."

Assuming they work as I described, the major differences would be the frequency/chance of taking identity hash codes and the need for the GC to call into during the system weak sweeping.

Any more thoughts?

Patch-set: 1