Update patch set 6
Patch Set 6:

(4 comments)

Ian is looking into the r4 issues in another CL.  Why don't we go ahead and commit this change *with* the new model of having suspend checks in leafs.  Once we gather some performance data on this part of the change, we can consider the r4 issue.

Patch-set: 6
Label: Verified=0