Update patch set 2
Patch Set 2:

Vladimir, do you by any chance have a design document for this work? I am especially interested in the design of the aliasing algorithm, because a thorough review is needed in order to make sure everything is functionally sound.

Do you hope for this algorithm to actually do the redundancy elimination? Or will it annotate in such a way that other algorithms (both ME and BE) can make use of the value number and aliasing information?

Patch-set: 2
Reviewer: Gerrit User 1013855 <1013855@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0