Update patch set 6
Patch Set 6:

Dmitry, looks good.  There are things that need to be done (by me) to generalize for all targets, and other places (such as GetArgMappingToPhysicalReg) which need to be adapted to no longer assume that all arguments are passed via core regs.  But, I'd rather get this CL committed so we can work from a common base.  If you can get this into a state that doesn't break existing targets and advances x86_64, then let's check it in.

Patch-set: 6
Label: Verified=0