Update patch set 2
Patch Set 2: Code-Review+1

Bill, when I first submitted the compiler temporary patch, I made it conform to the existing numbering scheme (as negative values). That was fine because the BE ended up using it to reserve space to help with method ptr and linking. However, we now believe that it makes sense to make the compiler temporaries positive when used in the middle-end. This is because data structures like bit vectors like positive values. This kind of change allow for infrastructure to not be aware of distinction between temps and VRs unless they are explicitly concerned. Because the change is minimal and no algorithms yet in the ME need temps, this change should be okay to let through :) Thanks so much!

Patch-set: 2
Label: Code-Review=+1