Update patch set 3
Patch Set 3:

The test does not need to go into jni_compiler_test, as it's not generic testing. This is specific to the assembler, and we do have a test for that one. I added the tests in PS3, and you will see that the test fails in 32b host mode without the changes. In the future, when there are other fixes in the assembler, please add a test, it is really not hard.

Patch-set: 3
Label: Verified=0