Update patch set 3
Patch Set 3:

Your hack is to say that an 'r' rex prefix is a 'b' prefix in special cases. That's not what I meant. I meant that if two instructions don't share the same encoding 'kind' then you need two encoding kinds. This is what we do with kRegReg and kRegRegStore.

Patch-set: 3