Update patch set 3
Patch Set 3:

We have the same problem with kReg vs kRegOpcode, there's a hack there where we've hacked ComputeSize and are passing the register in the wrong place (I added notes to the hacks). I would like to see this cleaned up. I think for the Emits we just need two forms (like kReg and kRegOpcode) but ComputeSize also needs cleaning up for the register form cases.

Patch-set: 3
Label: Verified=0