Update patch set 11
Patch Set 11:

We can enable compilation by default for x86_64 now, the JNI compiler patch is merged (Ian, Andreas, thanks for your help with test!).

We do not enable compilation of core.oat by this patch since there is a blocker issue for BTG reproducible when core.oat/boot.oat compilation is enabled. We are working on that issue and will provide a fix as a separate CL

The patch https://android-review.googlesource.com/#/c/97501/ is needed to be merged before this patch (without dependency on https://android-review.googlesource.com/#/c/97015/6).

The https://android-review.googlesource.com/#/c/97015/6 breaks the BTG (Boot-to-GUI) and needs some more testing.

Patch-set: 11