Update patch set 2
Patch Set 2:

(1 comment)

Thanks Brian for having taken a look.

 > if all tests are passing now, why do we want to proactively add a
 > mechanism to suppress failing tests.

I don't want people in the team get stuck because they add a test that fail on the optimizing compiler. They should submit their work and just add the test as a failing test for the compiler.

It does run now, but there are plenty of cases that could make it break. For example, it doesn't generate stack overflow checks, or suspend checks.

Patch-set: 2