Update patch set 5
Patch Set 5:

In the comparison

  mir_graph_->SRegToVReg(rl_dest.s_reg_low) == mir_graph_->SRegToVReg(rl_lhs.s_reg_low)

the rl_dest may not have an SSA register, i.e. rl_dest.s_reg_low == -1, and the call to SRegToVReg(-1) is invalid. We need to fix this CL to handle that situation and then we can resubmit.

Since the comparison also appears twice in int_x86.cc we should thoroughly review that usage and fix if needed.

Patch-set: 5