Update patch set 3
Patch Set 3:

(1 comment)

I'd prefer not to add the 'b' to the opcode in the case of mov but add it uniformly (so it benefits cmp, etc.). We can do this by appending b to the opcode when byte_operand is true. We can similarly add w, l  and q based on the prefixes.

Patch-set: 3
Label: Verified=0