Update patch set 2
Patch Set 2:

In order to be able to merge this, we need to fix all the improper instruction usages. I think there are quite a few (many possibly stem from usages of TargetReg()). I am therefore in favour of splitting this CL as you suggest. Feel free to do this yourself and separately commit/edit/refactor whatever you feel is appropriate. In a way, I pushed this as a WIP for this very reason. Also notice that it will be difficult for me to push anything new earlier than on Monday.

Patch-set: 2