Update patch set 2
Patch Set 2:

This loses ashmem names on the host with a ulimit set. The host names are of the form android-ashmem-<pid>-<8 char random number>, that is they don't contain meaningful names like "dalvik-heap". This change has no effect on the target.

To avoid the effect of this patch you can just run tests without a ulimit, which could be something we do in --dev or --gdb mode of run-test.

Having a time out would also be good, but any value would have to be in the minutes. If you are running tests on your machine that can mean minutes of a DOS on your use, whereas the filesize limit tends to cause tests fail fast - or avoid the worst aspect of the DOS that your disk is in constant use.

Patch-set: 2