Update patch set 1
Patch Set 1:

(6 comments)

Do you need both types of profiling, i.e. DexPC and Stack? The first one is a subset and could easily be generated if you choose a 0 bound when walking the stack.

Also, the current format breaks profile format compatibility. I.e. If you generate the profiles using Stacks then ProfileFile will not be able to parse them. Currently, if the profiler is active the compiler driver will try to load any profiles and it will always fail to do so.

Maybe it's a good idea to add some magic string/number on the first line of the profiler to classify and version the file format.

Patch-set: 1