Update patch set 1
Patch Set 1:

This is a really good idea Calin!  Indeed, DexPC is just a special case (#frames==0).  We might not even need a magic number, but represent profile data with something like:

[ctx]:method/pc/size/freq?

The only downside is that this format doesn't offer any smart compression of contexts, but that should be fine for contexts of 2--3, which is already enough to reconstruct precise callgraph.  What do you think?

Patch-set: 1
Reviewer: Gerrit User 1039674 <1039674@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0