Update patch set 1
Patch Set 1:

About the DexPC and Stack:

Yes, DexPC is a special case of bounded call stack with depth set to 1. However, the current implementation of collecting DexPC used a different data structure to maintain the information and the dumped file is more compressed. If we used some smart compression for Stack, it would be difficult for data aggregation among different runs. Therefore, I consider keeping both format and perform some pilot experiments to evaluate them in terms of efficiency and data size. If we figured out that the smart compression in our pilot experiments is unnecessary, we can remove the DexPC type.

Adding some magic string at the beginning of the profile file is a good idea. I will do it.

Patch-set: 1
Label: Verified=0