Update patch set 3
Patch Set 3:

I think there are things we should think about before jumping to a choice:

While two copies/compactions have more overhead than one, does the existing SemiSpace collector really cause jank/problems? At least the assumption so far seemed to be that the compaction cost at app switching time isn't really causing perceptible jank.

If the existing SemiSpace collector compacts better (that is, less RAM usage) in the background apps, would it be better in terms of the system total RAM usage over this new compaction scheme? For example, if you have a lot of background apps, which I believe is typical, a better compaction may make a great deal of difference in terms of the total RAM usage, and the above-mentioned extra copy/compaction overhead may be worth trading for.

Would delayed OOMs be always better? An app could slow down to a crawl when it's almost out of memory, due to back-to-back GC runs, where neither the app or the GC makes little progress. It is so called "GC thrashing" and may be less desirable than a straight OOM and a restart. (One technique may be to give up on compaction when the heap is fuller than a certain point.)

Given this, I think we'd need more data.

The "Ros2Ros" naming in the code isn't to my taste. As Mathieu said, I believe it would be desirable to generalize it to support both rosalloc and dlmalloc and position it as a nice peer/parallel to the SemiSpace compaction path.

Patch-set: 3
Label: Verified=0