Update patch set 1
Patch Set 1:

Not a big deal, but could you reorder the registers slightly and make xSELF x19 and xSUSPEND x20.  I anticipate that at some point we'll make Ian happy by getting rid of xSUSPEND so this way we won't have an odd hole in the callee saves usable by the compiler.

Patch-set: 1