Update patch set 1
Patch Set 1:

I did this work in a different tree and for that reason I am still waiting to build all of this for emulator. :) If it doesn't compile, I promise I will make it do that soon :) Please let me know if you have comments on refactoring of the frontend.

Patch-set: 1
Label: Verified=0