Update patch set 1
Patch Set 1: Code-Review+1

Thanks for the update and for commenting on the status of the optimisations that are left disabled. When you mention RefReg issue, are you referring to CL 98605? Patch set 3 and related question have not received any attention, yet. I have stopped working on this (as I would need an answer to the question I posed). I assume that you prefer to put this on hold. I suspect you may want to commit your promotion-patch first or that you may want to fix these issues yourself (after all this is not arm64-specific). If this is not the case, then let me know.

Patch-set: 1
Reviewer: Gerrit User 1037505 <1037505@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1