Update patch set 2
Patch Set 2:

The first patch will only to clean-up the use of temp registers so that I can modify the ABI later without crash anything.

Yes, this change on ABI will have both good and bad side. We will need to do some benchmark(it might be difficult at this moment, since we do not have any 64-bit benchmark apk with native codes) and tests to see if it really benefits us.

Patch-set: 2