Update patch set 2
Patch Set 2:

The Devil's advocate would say that Andreas' objection would apply to RegStorage as well.

In particular, one could unpack all the RegStorage arguments into their fundamental components. This would indeed allow us to check them, as - similarly to the extend/shift case - not all the possible values for a RegStorage are legal.

In fact, I don't think checking the shift/extend is a real problem. This can be done in EncodeExtend, in a dedicated function or in the assembler (similarly to what is done for RegStorage). In particular, I think that currently all the functions that are using EncodeExtend need to carry out the same check. Things will change when/if we will implement load/stores with extended register offset, as these instructions take a restricted set of extend expressions. For these, the best place to carry out the check is probably the assembler (with a dedicated kFmt, I guess).

On the other hand, using EncodeExtend and EncodeShift directly, you could define a single OpRegRegReg function, which could be used as this:

  OpRegRegReg(rs_x0, rs_x1, rs_x2, EncodeShift(...))
  OpRegRegReg(rs_x0, rs_x1, rs_x2, EncodeExtend(...))

Said that, this is - after all - mainly aesthetics and I am happy with your proposal as well :-)

Patch-set: 2
Reviewer: Gerrit User 1037505 <1037505@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0